Move core config functionality to its own module
