+++ homeassistant/core.py
@@ -3145,7 +3145,7 @@ class Config:
     async def async_update(self, **kwargs: Any) -> None:
         """Update the configuration from a dictionary."""
         # pylint: disable-next=import-outside-toplevel
-        from .config import (
+        from .core_config import (
             _raise_issue_if_historic_currency,
             _raise_issue_if_no_country,
         )
