+++ homeassistant/helpers/check_config.py
@@ -13,7 +13,6 @@ import voluptuous as vol
 from homeassistant import loader
 from homeassistant.config import (  # type: ignore[attr-defined]
     CONF_PACKAGES,
-    CORE_CONFIG_SCHEMA,
     YAML_CONFIG_FILE,
     config_per_platform,
     extract_domain_configs,
@@ -23,6 +22,7 @@ from homeassistant.config import (  # type: ignore[attr-defined]
     merge_packages_config,
 )
 from homeassistant.core import DOMAIN as HOMEASSISTANT_DOMAIN, HomeAssistant
+from homeassistant.core_config import CORE_CONFIG_SCHEMA
 from homeassistant.exceptions import HomeAssistantError
 from homeassistant.requirements import (
     RequirementsNotFound,
