+++ homeassistant/helpers/entity.py
@@ -21,7 +21,6 @@ from typing import TYPE_CHECKING, Any, Final, Literal, NotRequired, TypedDict, f
 from propcache import cached_property
 import voluptuous as vol
 
-from homeassistant.config import DATA_CUSTOMIZE
 from homeassistant.const import (
     ATTR_ASSUMED_STATE,
     ATTR_ATTRIBUTION,
@@ -49,6 +48,7 @@ from homeassistant.core import (
     get_hassjob_callable_job_type,
     get_release_channel,
 )
+from homeassistant.core_config import DATA_CUSTOMIZE
 from homeassistant.exceptions import (
     HomeAssistantError,
     InvalidStateError,
