+++ tests/components/google_assistant/test_helpers.py
@@ -15,8 +15,8 @@ from homeassistant.components.google_assistant.const import (
     STORE_GOOGLE_LOCAL_WEBHOOK_ID,
 )
 from homeassistant.components.matter import MatterDeviceInfo
-from homeassistant.config import async_process_ha_core_config
 from homeassistant.core import HomeAssistant, State
+from homeassistant.core_config import async_process_ha_core_config
 from homeassistant.helpers import device_registry as dr, entity_registry as er
 from homeassistant.setup import async_setup_component
 from homeassistant.util import dt as dt_util
