+++ tests/components/mailgun/test_init.py
@@ -8,9 +8,9 @@ import pytest
 
 from homeassistant import config_entries
 from homeassistant.components import mailgun, webhook
-from homeassistant.config import async_process_ha_core_config
 from homeassistant.const import CONF_API_KEY, CONF_DOMAIN
 from homeassistant.core import Event, HomeAssistant, callback
+from homeassistant.core_config import async_process_ha_core_config
 from homeassistant.data_entry_flow import FlowResultType
 from homeassistant.setup import async_setup_component
