+++ tests/components/rest/test_init.py
@@ -12,6 +12,7 @@ from homeassistant import config as hass_config
 from homeassistant.components.rest.const import DOMAIN
 from homeassistant.const import (
     ATTR_ENTITY_ID,
+    CONF_PACKAGES,
     SERVICE_RELOAD,
     STATE_UNAVAILABLE,
     UnitOfInformation,
@@ -468,7 +469,7 @@ async def test_config_schema_via_packages(hass: HomeAssistant) -> None:
         "pack_11": {"rest": {"resource": "http://url1"}},
         "pack_list": {"rest": [{"resource": "http://url2"}]},
     }
-    config = {HOMEASSISTANT_DOMAIN: {hass_config.CONF_PACKAGES: packages}}
+    config = {HOMEASSISTANT_DOMAIN: {CONF_PACKAGES: packages}}
     await hass_config.merge_packages_config(hass, config, packages)
 
     assert len(config) == 2
