+++ tests/components/webhook/test_init.py
@@ -9,8 +9,8 @@ from aiohttp.test_utils import TestClient
 import pytest
 
 from homeassistant.components import webhook
-from homeassistant.config import async_process_ha_core_config
 from homeassistant.core import HomeAssistant
+from homeassistant.core_config import async_process_ha_core_config
 from homeassistant.setup import async_setup_component
 
 from tests.typing import ClientSessionGenerator, WebSocketGenerator
